+++ test/doing_util_test.rb
@@ -34,9 +34,10 @@ class DoingUtilTest < Test::Unit::TestCase
   end
 
   def test_format_time
-    item = @wwid.content[@wwid.config['current_section']].items[0]
+    item = @wwid.content.in_section(@wwid.config['current_section'])[0]
+    distance = (item.end_date - item.date).to_i
     interval = @wwid.get_interval(item, formatted: false, record: false)
-    assert_equal(360, interval, 'Interval should match')
+    assert_equal(distance, interval, 'Interval should match')
     minutes = interval / 60 % 60
     res = @wwid.format_time(interval)
     assert_equal(minutes, res[2], 'Interval array should match')